Highest prio transformer: error cases tests
